Refactored classes into different packages
