+++ project/app/src/main/java/com/example/cmput301w21t25/activities_trials/ViewTrialActivity.java
@@ -1,4 +1,4 @@
-package com.example.cmput301w21t25.activities;
+package com.example.cmput301w21t25.activities_trials;
 
 import androidx.appcompat.app.AppCompatActivity;
